// src/views/demo-form/interfaces.ts
/**
 * Represents the schema definition for a demo form.
 */
export interface DemoFormSchema {
  /**
   * The value for the "name" field in the form.
   */
  name: string;

  /**
   * The value for the "lastName" field in the form.
   */
  lastName: string;

  /**
   * The contact details, including email and phone number.
   */
  contactDetails: {
    email: string;
    phoneNumber: string;
  };
}